added router, updated tests
