+++ src/App.vue
@@ -11,24 +11,6 @@ import Dashboard from '@/components/Dashboard';
 
 export default {
   name: 'App',
-  beforeMount() {
-    this.currentPage = this.getToken()
-      ? this.dynamicPages[1]
-      : this.dynamicPages[0];
-  },
-  data() {
-    return {
-      currentPage: null,
-      dynamicPages: ['app-principal-login', 'app-dashboard'],
-    };
-  },
-  methods: {
-    ...mapGetters(['getToken']),
-  },
-  components: {
-    'app-principal-login': PrincipalLogin,
-    Dashboard,
-  },
 };
 </script>
