+++ test/unit/specs/App.spec.js
@@ -1,47 +1,20 @@
 import { shallow, createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
-// import Vue from 'vue';
-import getters from '@/store/getters';
-import state from '@/store/initialStore';
+import Router from 'vue-router';
 
 import App from '@/App';
-import PrincipalLogin from '@/components/PrincipalLogin';
-import Dashboard from '@/components/Dashboard';
 
 const localVue = createLocalVue();
+const router = new Router();
 localVue.use(Vuex);
+localVue.use(Router);
 
 describe('App.vue ', () => {
-  // let getters;
-  let store;
-  let components;
-
-  beforeEach(() => {
-    store = new Vuex.Store({
-      state,
-      getters,
-    });
-
-    components = { 'app-principal-login': PrincipalLogin, 'app-dashboard': Dashboard };
-  });
-
-  it('should render login component', () => {
+  it('should render', () => {
     const wrapper = shallow(App, {
       localVue,
-      store,
-      stubs: components,
+      router,
     });
-
-    const h1 = wrapper.find('.login h1');
     expect(wrapper.isVueInstance()).toBeTruthy();
-    expect(h1.text()).toBe('Título de twitter');
-  });
-
-  it('should render dashboard component', () => {
-    const tokenizedStore = new Vuex.Store({ state: { token: 'defined' }, getters });
-    const wrapper = shallow(App, { localVue, stubs: components, store: tokenizedStore });
-
-    const h1 = wrapper.find('div h1');
-    expect(h1.text()).toBe('Dashboard');
   });
 });
